+++ Cargo.toml
@@ -114,6 +114,7 @@ optional = true
 test_util = { path = "components/test_util" }
 test_raftstore = { path = "components/test_raftstore" }
 test_storage = { path = "components/test_storage" }
+test_coprocessor = { path = "components/test_coprocessor" }
 criterion = "0.2"
 arrow = "0.10.0"
 
@@ -122,6 +123,7 @@ members = [
   "fuzz",
   "components/test_raftstore",
   "components/test_storage",
+  "components/test_coprocessor",
   "components/test_util"
 ]
