+++ components/test_coprocessor/Cargo.toml
@@ -1,0 +1,19 @@
+[package]
+name = "test_coprocessor"
+version = "0.0.1"
+publish = false
+
+[lib]
+path = "lib.rs"
+
+[dependencies]
+tikv = { path = "../../" }
+protobuf = "~2.0"
+test_storage = { path = "../test_storage" }
+futures = "0.1"
+
+[dependencies.kvproto]
+git = "https://github.com/pingcap/kvproto.git"
+
+[dependencies.tipb]
+git = "https://github.com/pingcap/tipb.git"
