+++ tests/coprocessor/test_select.rs
@@ -11,14 +11,10 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use std::collections::{BTreeMap, HashMap};
 use std::i64;
-use std::sync::atomic::{AtomicUsize, Ordering};
 use std::thread;
 use std::{cmp, mem};
 
-use futures::sync::{mpsc as futures_mpsc, oneshot};
-use futures::{Future, Stream};
 use protobuf::{Message, RepeatedField};
 
 use kvproto::coprocessor::{KeyRange, Request, Response};
@@ -27,38 +23,21 @@ use tipb::executor::{
     Aggregation, ExecType, Executor, IndexScan, Limit, Selection, TableScan, TopN,
 };
 use tipb::expression::{ByItem, Expr, ExprType, ScalarFuncSig};
-use tipb::schema::{self, ColumnInfo};
-use tipb::select::{Chunk, DAGRequest, EncodeType, SelectResponse, StreamResponse};
+use tipb::schema::ColumnInfo;
+use tipb::select::{Chunk, DAGRequest, EncodeType};
 
+use test_coprocessor::*;
 use test_storage::*;
 use tikv::coprocessor::codec::{datum, table, Datum};
 use tikv::coprocessor::*;
-use tikv::server::readpool::{self, ReadPool};
-use tikv::server::{Config, OnResponse};
-use tikv::storage::engine::{self, Engine, RocksEngine, TEMP_DIR};
-use tikv::storage::{self, Key, Mutation, ALL_CFS};
+use tikv::server::Config;
+use tikv::storage::engine::{self, RocksEngine, TEMP_DIR};
+use tikv::storage::ALL_CFS;
 use tikv::util::codec::number::*;
-use tikv::util::worker::{Builder as WorkerBuilder, FutureWorker, Worker};
+use tikv::util::worker::Worker;
 
 const FLAG_IGNORE_TRUNCATE: u64 = 1;
 const FLAG_TRUNCATE_AS_WARNING: u64 = 1 << 1;
-
-static ID_GENERATOR: AtomicUsize = AtomicUsize::new(1);
-
-const TYPE_VAR_CHAR: i32 = 1;
-const TYPE_LONG: i32 = 2;
-
-pub fn next_id() -> i64 {
-    ID_GENERATOR.fetch_add(1, Ordering::Relaxed) as i64
-}
-
-fn check_chunk_datum_count(chunks: &[Chunk], datum_limit: usize) {
-    let mut iter = chunks.iter();
-    let res = iter.any(|x| datum::decode(&mut x.get_rows_data()).unwrap().len() != datum_limit);
-    if res {
-        assert!(iter.next().is_none());
-    }
-}
 
 struct DAGChunkSpliter {
     chunks: Vec<Chunk>,
@@ -97,431 +76,12 @@ impl Iterator for DAGChunkSpliter {
     }
 }
 
-#[derive(Clone, Copy)]
-pub struct Column {
-    id: i64,
-    col_type: i32,
-    // negative means not a index key, 0 means primary key, positive means normal index key.
-    pub index: i64,
-    default_val: Option<i64>, // TODO: change it to Vec<u8> if other type value is needed for test.
-}
-
-struct ColumnBuilder {
-    col_type: i32,
-    index: i64,
-    default_val: Option<i64>,
-}
-
-impl ColumnBuilder {
-    fn new() -> ColumnBuilder {
-        ColumnBuilder {
-            col_type: TYPE_LONG,
-            index: -1,
-            default_val: None,
-        }
+fn check_chunk_datum_count(chunks: &[Chunk], datum_limit: usize) {
+    let mut iter = chunks.iter();
+    let res = iter.any(|x| datum::decode(&mut x.get_rows_data()).unwrap().len() != datum_limit);
+    if res {
+        assert!(iter.next().is_none());
     }
-
-    fn col_type(mut self, t: i32) -> ColumnBuilder {
-        self.col_type = t;
-        self
-    }
-
-    fn primary_key(mut self, b: bool) -> ColumnBuilder {
-        if b {
-            self.index = 0;
-        } else {
-            self.index = -1;
-        }
-        self
-    }
-
-    fn index_key(mut self, idx_id: i64) -> ColumnBuilder {
-        self.index = idx_id;
-        self
-    }
-
-    fn default(mut self, val: i64) -> ColumnBuilder {
-        self.default_val = Some(val);
-        self
-    }
-
-    fn build(self) -> Column {
-        Column {
-            id: next_id(),
-            col_type: self.col_type,
-            index: self.index,
-            default_val: self.default_val,
-        }
-    }
-}
-
-pub struct Table {
-    id: i64,
-    handle_id: i64,
-    cols: BTreeMap<i64, Column>,
-    idxs: BTreeMap<i64, Vec<i64>>,
-}
-
-impl Table {
-    fn get_table_info(&self) -> schema::TableInfo {
-        let mut tb_info = schema::TableInfo::new();
-        tb_info.set_table_id(self.id);
-        tb_info.set_columns(RepeatedField::from_vec(self.get_table_columns()));
-        tb_info
-    }
-
-    pub fn get_table_columns(&self) -> Vec<ColumnInfo> {
-        let mut tb_info = Vec::new();
-        for col in self.cols.values() {
-            let mut c_info = ColumnInfo::new();
-            c_info.set_column_id(col.id);
-            c_info.set_tp(col.col_type);
-            c_info.set_pk_handle(col.index == 0);
-            if let Some(dv) = col.default_val {
-                c_info.set_default_val(datum::encode_value(&[Datum::I64(dv)]).unwrap())
-            }
-            tb_info.push(c_info);
-        }
-        tb_info
-    }
-
-    fn get_index_info(&self, index: i64, store_handle: bool) -> schema::IndexInfo {
-        let mut idx_info = schema::IndexInfo::new();
-        idx_info.set_table_id(self.id);
-        idx_info.set_index_id(index);
-        let mut has_pk = false;
-        for col_id in &self.idxs[&index] {
-            let col = self.cols[col_id];
-            let mut c_info = ColumnInfo::new();
-            c_info.set_tp(col.col_type);
-            c_info.set_column_id(col.id);
-            if col.id == self.handle_id {
-                c_info.set_pk_handle(true);
-                has_pk = true
-            }
-            idx_info.mut_columns().push(c_info);
-        }
-        if !has_pk && store_handle {
-            let mut handle_info = ColumnInfo::new();
-            handle_info.set_tp(TYPE_LONG);
-            handle_info.set_column_id(-1);
-            handle_info.set_pk_handle(true);
-            idx_info.mut_columns().push(handle_info);
-        }
-        idx_info
-    }
-
-    pub fn get_select_range(&self) -> KeyRange {
-        let mut range = KeyRange::new();
-        range.set_start(table::encode_row_key(self.id, i64::MIN));
-        range.set_end(table::encode_row_key(self.id, i64::MAX));
-        range
-    }
-
-    pub fn get_index_range(&self, idx: i64) -> KeyRange {
-        let mut range = KeyRange::new();
-        let mut buf = Vec::with_capacity(8);
-        buf.encode_i64(i64::MIN).unwrap();
-        range.set_start(table::encode_index_seek_key(self.id, idx, &buf));
-        buf.clear();
-        buf.encode_i64(i64::MAX).unwrap();
-        range.set_end(table::encode_index_seek_key(self.id, idx, &buf));
-        range
-    }
-}
-
-struct TableBuilder {
-    handle_id: i64,
-    cols: BTreeMap<i64, Column>,
-}
-
-impl TableBuilder {
-    fn new() -> TableBuilder {
-        TableBuilder {
-            handle_id: -1,
-            cols: BTreeMap::new(),
-        }
-    }
-
-    fn add_col(mut self, col: Column) -> TableBuilder {
-        if col.index == 0 {
-            if self.handle_id > 0 {
-                self.handle_id = 0;
-            } else if self.handle_id < 0 {
-                // maybe need to check type.
-                self.handle_id = col.id;
-            }
-        }
-        self.cols.insert(col.id, col);
-        self
-    }
-
-    fn build(mut self) -> Table {
-        if self.handle_id <= 0 {
-            self.handle_id = next_id();
-        }
-        let mut idx = BTreeMap::new();
-        for (&id, col) in &self.cols {
-            if col.index < 0 {
-                continue;
-            }
-            let e = idx.entry(col.index).or_insert_with(Vec::new);
-            e.push(id);
-        }
-        for (id, val) in &mut idx {
-            if *id == 0 {
-                continue;
-            }
-            // TODO: support uniq index.
-            val.push(self.handle_id);
-        }
-        Table {
-            id: next_id(),
-            handle_id: self.handle_id,
-            cols: self.cols,
-            idxs: idx,
-        }
-    }
-}
-
-struct Insert<'a, E: Engine> {
-    store: &'a mut Store<E>,
-    table: &'a Table,
-    values: BTreeMap<i64, Datum>,
-}
-
-impl<'a, E: Engine> Insert<'a, E> {
-    fn new(store: &'a mut Store<E>, table: &'a Table) -> Self {
-        Insert {
-            store,
-            table,
-            values: BTreeMap::new(),
-        }
-    }
-
-    fn set(mut self, col: Column, value: Datum) -> Self {
-        assert!(self.table.cols.contains_key(&col.id));
-        self.values.insert(col.id, value);
-        self
-    }
-
-    fn execute(self) -> i64 {
-        self.execute_with_ctx(Context::new())
-    }
-
-    fn execute_with_ctx(self, ctx: Context) -> i64 {
-        let handle = self
-            .values
-            .get(&self.table.handle_id)
-            .cloned()
-            .unwrap_or_else(|| Datum::I64(next_id()));
-        let key = table::encode_row_key(self.table.id, handle.i64());
-        let ids: Vec<_> = self.values.keys().cloned().collect();
-        let values: Vec<_> = self.values.values().cloned().collect();
-        let value = table::encode_row(values, &ids).unwrap();
-        let mut kvs = vec![];
-        kvs.push((key, value));
-        for (&id, idxs) in &self.table.idxs {
-            let mut v: Vec<_> = idxs.iter().map(|id| self.values[id].clone()).collect();
-            v.push(handle.clone());
-            let encoded = datum::encode_key(&v).unwrap();
-            let idx_key = table::encode_index_seek_key(self.table.id, id, &encoded);
-            kvs.push((idx_key, vec![0]));
-        }
-        self.store.put(ctx, kvs);
-        handle.i64()
-    }
-}
-
-struct Delete<'a, E: Engine> {
-    store: &'a mut Store<E>,
-    table: &'a Table,
-}
-
-impl<'a, E: Engine> Delete<'a, E> {
-    fn new(store: &'a mut Store<E>, table: &'a Table) -> Self {
-        Delete { store, table }
-    }
-
-    fn execute(self, id: i64, row: Vec<Datum>) {
-        let mut values = HashMap::new();
-        for (&id, v) in self.table.cols.keys().zip(row) {
-            values.insert(id, v);
-        }
-        let key = table::encode_row_key(self.table.id, id);
-        let mut keys = vec![];
-        keys.push(key);
-        for (&idx_id, idx_cols) in &self.table.idxs {
-            let mut v: Vec<_> = idx_cols.iter().map(|id| values[id].clone()).collect();
-            v.push(Datum::I64(id));
-            let encoded = datum::encode_key(&v).unwrap();
-            let idx_key = table::encode_index_seek_key(self.table.id, idx_id, &encoded);
-            keys.push(idx_key);
-        }
-        self.store.delete(keys);
-    }
-}
-
-pub struct Store<E: Engine> {
-    store: SyncStorage<E>,
-    current_ts: u64,
-    handles: Vec<Vec<u8>>,
-}
-
-impl<E: Engine> Store<E> {
-    fn new(engine: E) -> Self {
-        let pd_worker = FutureWorker::new("test-future–worker");
-        let read_pool = ReadPool::new("readpool", &readpool::Config::default_for_test(), || {
-            || storage::ReadPoolContext::new(pd_worker.scheduler())
-        });
-        Self {
-            store: SyncStorage::from_engine(engine, &Default::default(), read_pool),
-            current_ts: 1,
-            handles: vec![],
-        }
-    }
-
-    pub fn get_engine(&self) -> E {
-        self.store.get_engine()
-    }
-
-    fn begin(&mut self) {
-        self.current_ts = next_id() as u64;
-        self.handles.clear();
-    }
-
-    fn insert_into<'a>(&'a mut self, table: &'a Table) -> Insert<'a, E> {
-        Insert::new(self, table)
-    }
-
-    fn put(&mut self, ctx: Context, mut kv: Vec<(Vec<u8>, Vec<u8>)>) {
-        self.handles.extend(kv.iter().map(|&(ref k, _)| k.clone()));
-        let pk = kv[0].0.clone();
-        let kv = kv
-            .drain(..)
-            .map(|(k, v)| Mutation::Put((Key::from_raw(&k), v)))
-            .collect();
-        self.store.prewrite(ctx, kv, pk, self.current_ts).unwrap();
-    }
-
-    fn delete_from<'a>(&'a mut self, table: &'a Table) -> Delete<'a, E> {
-        Delete::new(self, table)
-    }
-
-    fn delete(&mut self, mut keys: Vec<Vec<u8>>) {
-        self.handles.extend(keys.clone());
-        let pk = keys[0].clone();
-        let mutations = keys
-            .drain(..)
-            .map(|k| Mutation::Delete(Key::from_raw(&k)))
-            .collect();
-        self.store
-            .prewrite(Context::new(), mutations, pk, self.current_ts)
-            .unwrap();
-    }
-
-    fn commit(&mut self) {
-        self.commit_with_ctx(Context::new());
-    }
-
-    fn commit_with_ctx(&mut self, ctx: Context) {
-        let handles = self.handles.drain(..).map(|x| Key::from_raw(&x)).collect();
-        self.store
-            .commit(ctx, handles, self.current_ts, next_id() as u64)
-            .unwrap();
-    }
-}
-
-/// An example table for test purpose.
-pub struct ProductTable {
-    pub id: Column,
-    pub name: Column,
-    pub count: Column,
-    pub table: Table,
-}
-
-impl ProductTable {
-    pub fn new() -> ProductTable {
-        let id = ColumnBuilder::new()
-            .col_type(TYPE_LONG)
-            .primary_key(true)
-            .build();
-        let idx_id = next_id();
-        let name = ColumnBuilder::new()
-            .col_type(TYPE_VAR_CHAR)
-            .index_key(idx_id)
-            .build();
-        let count = ColumnBuilder::new()
-            .col_type(TYPE_LONG)
-            .index_key(idx_id)
-            .build();
-        let table = TableBuilder::new()
-            .add_col(id)
-            .add_col(name)
-            .add_col(count)
-            .build();
-
-        ProductTable {
-            id,
-            name,
-            count,
-            table,
-        }
-    }
-}
-
-fn init_data_with_engine_and_commit<E: Engine>(
-    ctx: Context,
-    engine: E,
-    tbl: &ProductTable,
-    vals: &[(i64, Option<&str>, i64)],
-    commit: bool,
-) -> (Store<E>, Worker<EndPointTask<E>>) {
-    init_data_with_details(ctx, engine, tbl, vals, commit, Config::default())
-}
-
-fn init_data_with_details<E: Engine>(
-    ctx: Context,
-    engine: E,
-    tbl: &ProductTable,
-    vals: &[(i64, Option<&str>, i64)],
-    commit: bool,
-    cfg: Config,
-) -> (Store<E>, Worker<EndPointTask<E>>) {
-    let mut store = Store::new(engine);
-
-    store.begin();
-    for &(id, name, count) in vals {
-        store
-            .insert_into(&tbl.table)
-            .set(tbl.id, Datum::I64(id))
-            .set(tbl.name, name.map(|s| s.as_bytes()).into())
-            .set(tbl.count, Datum::I64(count))
-            .execute_with_ctx(ctx.clone());
-    }
-    if commit {
-        store.commit_with_ctx(ctx);
-    }
-    let pd_worker = FutureWorker::new("test-pd-worker");
-    let pool = ReadPool::new("readpool", &readpool::Config::default_for_test(), || {
-        || ReadPoolContext::new(pd_worker.scheduler())
-    });
-    let mut end_point = WorkerBuilder::new("test select worker")
-        .batch_size(5)
-        .create();
-    let runner = EndPointHost::new(store.get_engine(), end_point.scheduler(), &cfg, pool);
-    end_point.start(runner).unwrap();
-
-    (store, end_point)
-}
-
-pub fn init_data_with_commit(
-    tbl: &ProductTable,
-    vals: &[(i64, Option<&str>, i64)],
-    commit: bool,
-) -> (Store<RocksEngine>, Worker<EndPointTask<RocksEngine>>) {
-    let engine = engine::new_local_engine(TEMP_DIR, ALL_CFS).unwrap();
-    init_data_with_engine_and_commit(Context::new(), engine, tbl, vals, commit)
 }
 
 // This function will create a Product table and initialize with the specified data.
@@ -1502,52 +1062,6 @@ fn test_reverse() {
     assert_eq!(row_count, 5);
 
     end_point.stop().unwrap().join().unwrap();
-}
-
-pub fn handle_request<E: Engine>(end_point: &Worker<EndPointTask<E>>, req: Request) -> Response {
-    let (tx, rx) = oneshot::channel();
-    let on_resp = OnResponse::Unary(tx);
-    let req = RequestTask::new(String::from("127.0.0.1"), req, on_resp, 100).unwrap();
-    end_point.schedule(EndPointTask::Request(req)).unwrap();
-    rx.wait().unwrap()
-}
-
-fn handle_select<E: Engine>(end_point: &Worker<EndPointTask<E>>, req: Request) -> SelectResponse {
-    let resp = handle_request(end_point, req);
-    assert!(!resp.get_data().is_empty(), "{:?}", resp);
-    let mut sel_resp = SelectResponse::new();
-    sel_resp.merge_from_bytes(resp.get_data()).unwrap();
-    sel_resp
-}
-
-fn handle_streaming_select<E: Engine, F>(
-    end_point: &Worker<EndPointTask<E>>,
-    req: Request,
-    mut check_range: F,
-) -> Vec<StreamResponse>
-where
-    F: FnMut(&Response) + Send + 'static,
-{
-    let (stream_tx, stream_rx) = futures_mpsc::channel(10);
-    let req = RequestTask::new(
-        String::from("127.0.0.1"),
-        req,
-        OnResponse::Streaming(stream_tx),
-        100,
-    ).unwrap();
-    end_point.schedule(EndPointTask::Request(req)).unwrap();
-    stream_rx
-        .wait()
-        .into_iter()
-        .map(|resp| {
-            let resp = resp.unwrap();
-            check_range(&resp);
-            assert!(!resp.get_data().is_empty());
-            let mut stream_resp = StreamResponse::new();
-            stream_resp.merge_from_bytes(resp.get_data()).unwrap();
-            stream_resp
-        })
-        .collect()
 }
 
 #[test]
